backend/requirements.txt
# IntelliRelief Backend Dependencies
# Python 3.11+

# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==3.2.2

# ============================================
# REDIS & CACHING
# ============================================
redis==5.0.1
aioredis==2.0.1

# ============================================
# HTTP CLIENT
# ============================================
httpx==0.27.0
aiohttp==3.9.3

# ============================================
# WEBSOCKET
# ============================================
python-socketio==5.11.1
websockets==12.0

# ============================================
# DATA VALIDATION & SERIALIZATION
# ============================================
python-dateutil==2.8.2
email-validator==2.1.0

# ============================================
# LOGGING & MONITORING
# ============================================
loguru==0.7.2

# ============================================
# TESTING
# ============================================
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.27.0
faker==22.6.0

# ============================================
# DEVELOPMENT TOOLS
# ============================================
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# ============================================
# EXTERNAL INTEGRATIONS
# ============================================
# requests kept for any non-async scripts; httpx is used for all async calls
requests==2.31.0

# For voice-to-text (local Whisper model, ~75 MB for 'tiny')
faster-whisper>=1.0.1

# For email notifications
aiosmtplib==3.0.1

# For file handling
python-magic==0.4.27

# For task scheduling (background jobs)
celery==5.3.6

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.1